% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_Met_objects.R
\name{all_Met_objects}
\alias{all_Met_objects}
\title{Returns the number of available Open Access objects in the Met Collection and their objectIDs.}
\usage{
all_Met_objects(limit = FALSE, include_IDs = FALSE)
}
\arguments{
\item{limit}{Limits the number of objectIDs returned by the function.}

\item{include_IDs}{If set to TRUE, the function will return objectIDs, observing the limit argument. Defaults to FALSE.}
}
\description{
Returns the number of available Open Access objects in the Met Collection and their objectIDs.
}
