% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_department_IDs.R
\name{get_department_IDs}
\alias{get_department_IDs}
\title{Get department IDs}
\usage{
get_department_IDs()
}
\description{
This function returns information on the departments currently at the Metropolitan Museum. Running this function will output a data.frame of the names and DepartmentIDs of the current departments.
}
