% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Met_object_info.R
\name{get_Met_object_info}
\alias{get_Met_object_info}
\title{Searches by objectID and eturns all available data.}
\usage{
get_Met_object_info(objectID = "", download_image = FALSE)
}
\arguments{
\item{objectID}{The Metropolitan Museum's given ID for an object.}

\item{download_image}{If TRUE, downloads primary image files the object in the query. Defaults to FALSE.}
}
\description{
Searches by objectID and eturns all available data.
}
