% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_Met_objects.R
\name{search_Met_objects}
\alias{search_Met_objects}
\title{Return object data based on search paramaters}
\usage{
search_Met_objects(
  q = "",
  departmentID = "",
  isHighlight = FALSE,
  isOnView = FALSE,
  artistOrCulture = FALSE,
  medium = "",
  hasImages = FALSE,
  geoLocation = "",
  dateBegin = "",
  dateEnd = "",
  limit = 5,
  download_images = FALSE
)
}
\arguments{
\item{q}{Search query. Returns a listing of all Object IDs for objects that contain the search query anywhere within the object’s data.}

\item{departmentID}{Returns objects that are a part of a specific department. Use `get_department_IDs()` function to explore available departments and their associated IDs.}

\item{isHighlight}{If TRUE, filters for objects that match the query and are designated as highlights. Highlights are selected works of art from The Met Museum’s permanent collection representing different cultures and time periods.}

\item{isOnView}{If TRUE, filters for objects that match the query and are on view in the museum.}

\item{artistOrCulture}{If TRUE, filters for objects that match the query, specifically searching against the artist name or culture field for objects.}

\item{medium}{Returns objects that match the query and are of the specified medium or object type. Examples include: Ceramics, Furniture, Paintings, Sculpture, Textiles, etc.}

\item{hasImages}{If TRUE, filters for objects that match the query and have images.}

\item{geoLocation}{Returns objects that match the query and the specified geographic location. Examples include: Europe, France, Paris, China, New York, etc.}

\item{dateBegin}{Returns objects that match the query and fall between the dateBegin and dateEnd parameters. Must also be used with dateEnd argument.}

\item{dateEnd}{Returns objects that match the query and fall between the dateBegin and dateEnd parameters. Must also be used with dateBegin argument.}

\item{limit}{Limits the number of objects that have data included in the output. Defaults to 5.}

\item{download_images}{If TRUE, downloads primary image files of each object in the query. Observes limit paramater.  Defaults to FALSE.}
}
\description{
This function searches for all objects in the Met's Open Access database that match the search parameters. The user can also use this function to download data on a number of these objects (5 by default) and download the images (off by default). Some of these arguments come directly from the parameters in the Met's API. In these cases, the same descriptions are used.
}
